Changed isa.h file and added isa.c file
